+++ LabClothingCollectionAPI/Models/CollectionDto.cs
@@ -8,7 +8,7 @@ namespace LabClothingCollectionAPI.Models
         public string Name { get; set; } = string.Empty;
         public string Owner { get; set; } = string.Empty;
         public string Brand { get; set; } = string.Empty;
-        public decimal Budget { get; set; }
+        public double Budget { get; set; }
         public DateTime ReleaseDate { get; set; }
         public Season Season { get; set; }
         public Status Status { get; set; }
